+++ src/pages/Funcionario/index.tsx
@@ -5,6 +5,8 @@ import api from "../../services/api";
 import { toast } from "react-toastify";
 import { AiOutlineUpload } from "react-icons/ai";
 import { MdDeleteForever } from "react-icons/md";
+import { confirmAlert } from "react-confirm-alert";
+import 'react-confirm-alert/src/react-confirm-alert.css'
 
 import "./styles.css";
 import "../../components/Table/styles.css";
@@ -121,9 +123,27 @@ function FuncionarioPage(){
 
     async function excluir(id:number){
         try{
-            await api.delete(`funcionario/${id}`)
-            getFuncionario();
-            toast.success("Funcionario Excluído com sucesso");
+            // await api.delete(`funcionario/${id}`)
+            // getFuncionario();
+            // toast.success("Funcionario Excluído com sucesso");
+            confirmAlert({
+                title: 'Confirmar ação',
+                message: 'Tem certeza que deseja desativar o funcionário?',
+                buttons: [
+                    {
+                        label: 'Sim',
+                        onClick: ()=> {
+                            api.delete(`funcionario/${id}`)
+                            getFuncionario();
+                            toast.success("Funcionario Excluído com sucesso");
+                        }
+                    },
+                    {
+                        label: 'Não',
+                        onClick: ()=>{}
+                    }
+                ]
+            });
         } catch(err){
             toast.error("Funcionario não pode ser exluído");
         }
